Reports REST API tests
